Remove unnecessary variable and else-if #cosmetic
Recommend --word-diff-regex='\w+|.' to view diff.

I first noticed that we were doing isObject() twice if an object is
passed as `def` in the first place, and in simplifying that I realized
we didn't even need `extensions` in the first place.

`for (var ext in def)` didn't make much sense so I changed the name of
the iteration variable too.